tut_assert: use noreturn function attribute
These functions never return. They throw an exception. Add the C++ 11
noreturn function attribute to help compiler.

For the following code the compiler

    int f() {
        tut::fail("hello");
    }

emits a warning:

    In function ‘int f()’:
    warning: no return statement in function returning non-void [-Wreturn-type]

But this warning is a false postive. The code is actually correct,
because the function never returns normally.

With this patch the compiler does not produce this false-postive.